+++ modules/main/src/main/java/com/nisovin/shopkeepers/SKShopkeepersPlugin.java
@@ -26,6 +26,7 @@ import com.nisovin.shopkeepers.commands.Commands;
 import com.nisovin.shopkeepers.compat.MC_1_17;
 import com.nisovin.shopkeepers.compat.MC_1_19;
 import com.nisovin.shopkeepers.compat.MC_1_20;
+import com.nisovin.shopkeepers.compat.MC_1_20_6;
 import com.nisovin.shopkeepers.compat.NMSManager;
 import com.nisovin.shopkeepers.compat.ServerAssumptionsTest;
 import com.nisovin.shopkeepers.config.Settings;
@@ -317,6 +318,7 @@ public class SKShopkeepersPlugin extends JavaPlugin implements InternalShopkeepe
 		MC_1_17.init();
 		MC_1_19.init();
 		MC_1_20.init();
+		MC_1_20_6.init();
 
 		// Inform about Spigot exclusive features:
 		if (SpigotFeatures.isSpigotAvailable()) {