+++ modules/main/src/main/java/com/nisovin/shopkeepers/util/data/serialization/bukkit/NamespacedKeySerializers.java
@@ -1,0 +1,43 @@
+package com.nisovin.shopkeepers.util.data.serialization.bukkit;
+
+import org.bukkit.NamespacedKey;
+import org.checkerframework.checker.nullness.qual.NonNull;
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+import com.nisovin.shopkeepers.util.data.serialization.DataSerializer;
+import com.nisovin.shopkeepers.util.data.serialization.InvalidDataException;
+import com.nisovin.shopkeepers.util.data.serialization.java.StringSerializers;
+import com.nisovin.shopkeepers.util.java.Validate;
+
+/**
+ * Default {@link DataSerializer}s for {@link NamespacedKey} values.
+ */
+public final class NamespacedKeySerializers {
+
+	private static abstract class NamespacedKeySerializer implements DataSerializer<@NonNull NamespacedKey> {
+
+		@Override
+		public @Nullable Object serialize(NamespacedKey value) {
+			Validate.notNull(value, "value is null");
+			return value.toString();
+		}
+	}
+
+	/**
+	 * A {@link DataSerializer} for {@link NamespacedKey} values.
+	 */
+	public static final DataSerializer<@NonNull NamespacedKey> DEFAULT = new NamespacedKeySerializer() {
+		@Override
+		public NamespacedKey deserialize(Object data) throws InvalidDataException {
+			String keyString = StringSerializers.STRICT.deserialize(data);
+			@Nullable NamespacedKey key = NamespacedKey.fromString(keyString);
+			if (key == null) {
+				throw new InvalidDataException("Invalid namespaced key: '" + keyString + "'");
+			}
+			return key;
+		}
+	};
+
+	private NamespacedKeySerializers() {
+	}
+}